project/requirements.txt
asgiref==3.8.1
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.8
inflection==0.5.1
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==74.0.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
